+++ packages/sandbox/src/bin/cli.rs
@@ -2,7 +2,11 @@ use clap::{Args, Parser, Subcommand};
 use cmux_sandbox::models::{
     CreateSandboxRequest, EnvVar, ExecRequest, ExecResponse, SandboxSummary,
 };
-use cmux_sandbox::{AcpProvider, DEFAULT_HTTP_PORT};
+use cmux_sandbox::{
+    auth_files::{upload_auth_files, AUTH_FILES},
+    AcpProvider,
+    DEFAULT_HTTP_PORT,
+};
 use crossterm::terminal::{disable_raw_mode, enable_raw_mode};
 use futures::{SinkExt, StreamExt};
 use ignore::WalkBuilder;
@@ -121,115 +125,8 @@ enum AuthCommand {
     Status,
 }
 
-struct AuthFileDef {
-    name: &'static str,
-    host_path: &'static str,    // Relative to HOME
-    sandbox_path: &'static str, // Absolute in sandbox
-}
-
 const ENV_CMUX_NO_ATTACH: &str = "CMUX_NO_ATTACH";
 const ENV_CMUX_FORCE_ATTACH: &str = "CMUX_FORCE_ATTACH";
-
-const AUTH_FILES: &[AuthFileDef] = &[
-    AuthFileDef {
-        name: "Claude Config",
-        host_path: ".claude.json",
-        sandbox_path: "/root/.claude.json",
-    },
-    AuthFileDef {
-        name: "Codex Auth",
-        host_path: ".codex/auth.json",
-        sandbox_path: "/root/.codex/auth.json",
-    },
-    AuthFileDef {
-        name: "Codex Instructions",
-        host_path: ".codex/instructions.md",
-        sandbox_path: "/root/.codex/instructions.md",
-    },
-    AuthFileDef {
-        name: "Codex Config",
-        host_path: ".codex/config.toml",
-        sandbox_path: "/root/.codex/config.toml",
-    },
-    AuthFileDef {
-        name: "Gemini Settings",
-        host_path: ".gemini/settings.json",
-        sandbox_path: "/root/.gemini/settings.json",
-    },
-    AuthFileDef {
-        name: "Gemini OAuth",
-        host_path: ".gemini/oauth_creds.json",
-        sandbox_path: "/root/.gemini/oauth_creds.json",
-    },
-    AuthFileDef {
-        name: "Gemini MCP Tokens",
-        host_path: ".gemini/mcp-oauth-tokens.json",
-        sandbox_path: "/root/.gemini/mcp-oauth-tokens.json",
-    },
-    AuthFileDef {
-        name: "Gemini Google Accounts",
-        host_path: ".gemini/google_accounts.json",
-        sandbox_path: "/root/.gemini/google_accounts.json",
-    },
-    AuthFileDef {
-        name: "Gemini Account ID",
-        host_path: ".gemini/google_account_id",
-        sandbox_path: "/root/.gemini/google_account_id",
-    },
-    AuthFileDef {
-        name: "Gemini Install ID",
-        host_path: ".gemini/installation_id",
-        sandbox_path: "/root/.gemini/installation_id",
-    },
-    AuthFileDef {
-        name: "Gemini User ID",
-        host_path: ".gemini/user_id",
-        sandbox_path: "/root/.gemini/user_id",
-    },
-    AuthFileDef {
-        name: "Gemini Env",
-        host_path: ".gemini/.env",
-        sandbox_path: "/root/.gemini/.env",
-    },
-    AuthFileDef {
-        name: "Env",
-        host_path: ".env",
-        sandbox_path: "/root/.env",
-    },
-    AuthFileDef {
-        name: "OpenCode Auth",
-        host_path: ".local/share/opencode/auth.json",
-        sandbox_path: "/root/.local/share/opencode/auth.json",
-    },
-    // Amp
-    AuthFileDef {
-        name: "Amp Settings",
-        host_path: ".config/amp/settings.json",
-        sandbox_path: "/root/.config/amp/settings.json",
-    },
-    AuthFileDef {
-        name: "Amp Secrets",
-        host_path: ".local/share/amp/secrets.json",
-        sandbox_path: "/root/.local/share/amp/secrets.json",
-    },
-    // Cursor
-    AuthFileDef {
-        name: "Cursor CLI Config",
-        host_path: ".cursor/cli-config.json",
-        sandbox_path: "/root/.cursor/cli-config.json",
-    },
-    AuthFileDef {
-        name: "Cursor Auth",
-        host_path: ".config/cursor/auth.json",
-        sandbox_path: "/root/.config/cursor/auth.json",
-    },
-    // Qwen
-    AuthFileDef {
-        name: "Qwen Settings",
-        host_path: ".qwen/settings.json",
-        sandbox_path: "/root/.qwen/settings.json",
-    },
-];
 
 #[derive(Subcommand, Debug)]
 enum SandboxCommand {
@@ -1565,109 +1462,6 @@ async fn upload_workspace_directory(
             response.status()
         ));
     }
-    Ok(())
-}
-
-async fn upload_auth_files(
-    client: &Client,
-    base_url: &str,
-    id: &str,
-    log_progress: bool,
-) -> anyhow::Result<()> {
-    let home = std::env::var("HOME").unwrap_or_else(|_| ".".to_string());
-    let home_path = PathBuf::from(home);
-    let mut files_to_upload = Vec::new();
-
-    for def in AUTH_FILES {
-        let path = home_path.join(def.host_path);
-        if path.exists() {
-            files_to_upload.push((path, def.sandbox_path));
-        }
-    }
-
-    if files_to_upload.is_empty() {
-        return Ok(());
-    }
-
-    if log_progress {
-        eprintln!("Uploading {} auth files...", files_to_upload.len());
-    }
-
-    // Create a temporary directory structure for the tar
-    // We'll upload to /workspace/__cmux_auth_temp and then move
-    let (tx, rx) = tokio::sync::mpsc::channel::<Result<Vec<u8>, std::io::Error>>(10);
-
-    tokio::task::spawn_blocking(move || {
-        let writer = ChunkedWriter { sender: tx.clone() };
-        let mut tar = Builder::new(writer);
-
-        let temp_dir_name = "__cmux_auth_temp";
-
-        for (host_path, sandbox_path_str) in files_to_upload {
-            // sandbox_path is absolute, e.g. /root/.codex/auth.json
-            // We want to put it in the tar as __cmux_auth_temp/root/.codex/auth.json
-            // But tar paths shouldn't start with / usually if we want them relative to extraction point
-            // The server extracts to /workspace.
-            // So if we put it as "__cmux_auth_temp/...", it extracts to /workspace/__cmux_auth_temp/...
-
-            // We need to strip the leading / from sandbox_path for the tar entry
-            let sandbox_path = Path::new(sandbox_path_str);
-            let rel_sandbox_path = sandbox_path.strip_prefix("/").unwrap_or(sandbox_path);
-
-            let tar_path = Path::new(temp_dir_name).join(rel_sandbox_path);
-
-            if let Err(e) = tar.append_path_with_name(&host_path, &tar_path) {
-                let _ = tx.blocking_send(Err(e));
-                return;
-            }
-        }
-
-        if let Err(e) = tar.finish() {
-            let _ = tx.blocking_send(Err(e));
-        }
-    });
-
-    let body = reqwest::Body::wrap_stream(futures::stream::unfold(rx, |mut rx| async move {
-        rx.recv().await.map(|msg| (msg, rx))
-    }));
-
-    let url = format!("{}/sandboxes/{}/files", base_url.trim_end_matches('/'), id);
-    let response = client.post(url).body(body).send().await?;
-    if !response.status().is_success() {
-        return Err(anyhow::anyhow!(
-            "Failed to upload auth files: {}",
-            response.status()
-        ));
-    }
-
-    // Now move files from temp dir to final destination
-    // We uploaded to /workspace/__cmux_auth_temp/...
-    // We want to move contents of /workspace/__cmux_auth_temp/root/* to /root/
-    // and cleanup
-
-    let move_script = r#"
-        if [ -d /workspace/__cmux_auth_temp ]; then
-            cp -r /workspace/__cmux_auth_temp/root/. /root/ 2>/dev/null || true
-            rm -rf /workspace/__cmux_auth_temp
-        fi
-    "#;
-
-    let exec_body = ExecRequest {
-        command: vec!["/bin/sh".into(), "-c".into(), move_script.into()],
-        workdir: None,
-        env: Vec::new(),
-    };
-
-    let exec_url = format!("{}/sandboxes/{}/exec", base_url.trim_end_matches('/'), id);
-    let exec_response = client.post(exec_url).json(&exec_body).send().await?;
-
-    if !exec_response.status().is_success() {
-        return Err(anyhow::anyhow!(
-            "Failed to execute move script: {}",
-            exec_response.status()
-        ));
-    }
-
     Ok(())
 }
